--- LCDScreenOutput_3_18_21/BluetoothSetSend.h
@@ -1,14 +1,0 @@
-#ifndef BLUETOOTHSETSEND_H
-#define BLUETOOTHSETSEND_H
-
-void percentageStatus(int inc);
-
-void intesityStatus(int dataPoint);
-
-void bluetoothStatus();
-
-void bluetoothSetup();
-
-void printVal (char string[] , float data);
-
-#endif
