+++ LCDScreenOutput_3_18_21/LCDScreenOutput/ScreenDraw.h
@@ -1,0 +1,18 @@
+#ifndef SCREENDRAW_H
+#define SCREENDRAW_H
+
+void screenSetup();
+
+void drawingF(int Cx, int Cy, int R, int G, int B, int textSize);
+
+void percentageOutput(int i);
+
+void printWarning(); // Raven 3/17/21
+
+void printUV(); // Raven 3/6/21
+
+void graphUV(); // Raven 3/6/21
+
+void read_y2(); // Raven 3/12/21
+
+#endif
