+++ LCDScreenOutput_3_23_21/LCDScreenOutput/BluetoothSetSend.h
@@ -1,0 +1,14 @@
+#ifndef BLUETOOTHSETSEND_H
+#define BLUETOOTHSETSEND_H
+
+void percentageStatus(int inc);
+
+void intesityStatus(int dataPoint);
+
+void bluetoothStatus();
+
+void bluetoothSetup();
+
+void printVal (char string[] , float data);
+
+#endif
