+++ LCDScreenOutput_R_3_12_2021/LCDScreenOutput/LCDScreenOutput.ino
@@ -1,0 +1,43 @@
+
+#include "SPI.h"
+#include "ScreenDraw.h"
+#include "BluetoothSetSend.h"
+#include "UVSensorSetSend.h"
+
+int inc = 0;
+int dataPoint = 0;
+int timeVal = 0;
+
+
+
+
+void setup() {
+  screenSetup();
+  bluetoothSetup();
+  uvSensorSetup();
+}
+
+void loop() {
+
+  bluetoothStatus();
+  percentageStatus(inc);
+  intesityStatus(dataPoint);
+  uvSensorStatus();
+
+  // Raven 3/6/21
+  printUV();
+  graphUV();
+  
+  // incrementing percentage  
+  if(inc < 100)
+    inc = inc + 1;
+  else
+    inc = 0;
+
+  if(dataPoint < 50)
+    dataPoint = dataPoint + 1;
+  else
+    dataPoint = 0;  
+    
+  delay(100);
+}
